My Jetpack: fix multi-site availability check (#44710)
* Fix multisite availability logic for products and modules

The condition was previously inverted, causing the message “Not available on multisite” to be shown for products/modules not in the restriction list, instead of those that are in it.

* changelog

* Update changelog

* Suppress backup failure notice for multisite restriction

Do not show the red error bubble when the backup status is 'multisite_not_supported', as this is not a failure but an expected limitation.

* Hide multisite-unsupported products from card display

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/16883953386

Upstream-Ref: Automattic/jetpack@b87ff5911ffda9936c7bdd77740a870ab2db4b54